Refactoring: collapsing the two ClientRunner types into one class now called ClientRunner.java, we also do away with the interface by that name. Also removed unused param passing the context to a specific executor.
